Reworked recipe data structure
There were some issues regarding the original recipe data structure. As a result, it was reworked to maintain simplicity and an optimized run time. The recipes can now be printed in the chef recipe manager page, and they can be selected using radio buttons. Lastly, an executable file was created to allow ease in running the project for future viewers.